% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{cond_predict}
\alias{cond_predict}
\title{GP conditional predictions}
\usage{
cond_predict(
  y,
  x,
  x_pred,
  cov,
  ...,
  reps = 1000,
  diag_adj = 1e-06,
  cov_f_o = cov,
  cov_f_p = cov,
  cov_f_po = cov
)
}
\arguments{
\item{x}{Vector or matrix of fitted locations (values).}

\item{x_pred}{Vector or matrix of prediction locations (values).}

\item{cov}{Function. Covariance function that takes a distance matrix as input.}

\item{...}{Parameter values used by \code{cov}.}

\item{reps}{Numeric. Number of samples to draw from conditional distribution.}

\item{diag_adj}{Numeric. Small value that is added covariance diagonal to avoid
numerical singularity issues.}

\item{cov_f_o}{Function. Calculates the cross covariance matrix between the prediction and observed locations.}

\item{cov_f_p}{Function. Calculates the covariance matrix for the prediction locations.}
}
\value{
A matrix of size \code{n_pred} x \code{reps}.
}
\description{
Given a covariance function and parameters this samples from
the conditional MVN of y_pred | y
}
